ml/requirements_ml.txt
torch
torchvision
torchaudio
opencv-python
tqdm
pillow
kagglehub
scikit-learn
pytorch-grad-cam
numpy
